SpreadsheetUI : Support add/remove for StringVectorData drag and drop
This follows the pattern established in PathFilterUI, where <kbd>Shift</kbd> is used to add to the current list and <kbd>Control</kbd> is used to remove from it.

My first attempt at this involved implementing the add/remove behaviour in _ClipboardAlgo.py, passing a mode argument down through the `pasteCells()/ValueAdaptor.set()/ValueAdaptor._set()` chain. This felt a bit cumbersome given the amount of code it affected, and the fact that I could only ever imagine StringVectorDataPlugValueAdaptor using it. So instead I implemented what we need more locally in `_PlugTableView.__dropData()`.

Fixes #3894.