MessageWidget : No longer trigger UI updates while handling messages
This means that we will no longer update message logs in the UI while a slow process is running on the UI thread.  If you have a slow process that you want to receive updates from, you should run it on a background thread ( perhaps using BackgroundTaskDialogue ).  If you have old code that runs something slow on the UI thread, it will still run, but you will only see messages from it once it finishes.