ViewportGadget : Add `setTumblingEnabled()` and `getTumblingEnabled()`
Unlike `Viewport::m_cameraEditable` this is stored in the CameraController, to preserve ABI compatibility. It is checked in `CameraController::tumble()` rather than anywhere else because this is the simplest implementation, and because we want to swallow tumbling events even when disabled (so we don't get unexpected fall-through to other event handlers). As noted in the comments for `CameraController::cameraMotionType()`, the separation of concerns between CameraController and ViewportGadget is already somewhat eroded. If we were to refactor this, it seems that the camera movement code would be moved into 2DViewport and 3DViewport subclasses, and CameraController would be no more.