ViewportGadget : Modify orthographic aperture in dolly and frame
Adjusting the aperture to zoom in and out is what users intuitively expect from a "dolly" operation; it changes the projected size of what they're looking at, in a similar way to a conventional dolly for a perspective camera. It matches Maya behaviour. And it avoids situations where you are moving the camera further and further back in space with literally no visual feedback in the viewport.

Likewise, framing is much more useful if the aperture is adjusted to encompass everything you want to look at.

Alas, this new implementation is only useful for the newly added "free" cameras in the SceneView, and not for look-through cameras when manipulated by the CameraTool. We currently have no way to mirror the adjusted aperture into a graph edit, so the CameraTool disables dollying for orthographic cameras. This is an improvement on previous behaviour though, since you are no longer moving cameras backwards and forwards without feedback.

The use of `getDollyingEnabled()` to control the behaviour of `frame()` seems reasonable in the sense that it gives us the behaviour we need in practice. But it raises a question : "should `frame( viewDirection )` also respect `getTumblingEnabled()`?". That would seem to be logical, but I'm not sure it's useful, or relevant in practice, so I haven't gone there.