ValuePlug : Traverse `sourcePlug` across inherited classes
This is needed by `FilePathPlug`, which derives from `StringPlug`.
Without it, the case tested in
`GafferTest.FilePathPlugTest.testSubstitutionsFromExpressionInput`
fails. When the `FilePathPlug` evaluates `sourcePlug`, it never gets to
the expression's output `StringPlug` because of their `typeId` did not
match. And because the `FilePathPlug` being evaluated is an `Out` plug,
substitutions are never applied.

By changing the test in `sourcePlug()` to traverse across derived plug
types, the `StringPlug` with direction `In` makes the substitution and
the test passes.