ViewportGadget : Normalise CameraController's data representation
In `setCamera()` we were duplicating some camera state (resolution, clipping, aperture) into separate member variables for editing, and then putting it back again in `getCamera()`. This made `m_sourceCamera` an unreliable source of information in the intervening time. We had dealt with this until now by passing the film fit and resolution explicitly to all `m_sourceCamera->frustum()` calls, overriding whatever resolution was in the camera.

This approach failed when introducing aperture edits, because the aperture needs to be taken into account by `frustum()`. This introduced various bugs including tracking at the wrong speed and projection not working so that transform handles were drawn too thick. By keeping the camera in a valid state at all times, we can rely on the simpler argument-less `frustum()` call and all is well.