SceneView : Reposition default orthographic cameras
The original intention was to create a camera where the clipping planes were likely sufficient to show the entire scene, no matter the origin of the camera. But this was inconvenient when framing an object inside a room : the framing code would move the camera inside the walls, but the negative near clip would mean you still saw the outer wall and not the object.