GraphGadget : Remove redundant `dirty( Render )` call
This was originally added in e6445c9, apparently to cause a re-render if the GraphGadget was displaying the focus node. Logically this would be the responsibility of the FocusGadget itself, but the comment implies we're doing it here as an optimisation (although FocusGadget does so much signal handling that one more probably wouldn't hurt).

Later, f1b0749 added a call to `dirtyActive()` which in turn calls `dirty( Render )` to reflect the needed active-highlighting change (which _is_ the responsibility of the GraphGadget). So we can now remove the first `dirty()` call and rely on this second one alone.