Plug : Remove child outputs in `removeOutputs()`
This fixes #4614, where an unexpected remaining connection broke the updating of expressions when the parent of an input plug was removed. Our intention is that when a plug is removed, all its connections are also removed, so that it has no links to the remaining node graph. We were doing this correctly for input connections (including into children), but not for output connections from children.

This could have been fixed more locally by placing the child recursion in `parentChanging()` rather than in the public `removeOutputs()` method. In some ways its riskier to do it in `removeOutputs()`, but I believe it is the correct behaviour given that `setInput()` already recurses. This is backed up by examining the current uses of `removeOutputs()` in Gaffer - there are only two, and in both cases the removal of child outputs is desired.