CameraTool : Don't edit parent locations
When the camera came from a cache and therefore didn't have an editable transform, the `TransformTool::Selection` would "helpfully" make an edit to the root locations of the scene. This meant that attempts at camera movement moved the entire scene as well as the camera. Since the relative position of the two remained fixed, there was no feedback in the Viewer as to what was happening (unless you looked at the gnomon) and only later did it become apparent what a mess has been made. I fell into this trap repeatedly yesterday until I figured out what was going on.