config/ie : Better search for the corresponding registry entry
The previous logic would fail if we ever needed a new entry for the same
compatibility version, which happens to be the case for 0.61, where
support for Arnold 7.1.1.0 was added by the minor version 0.61.3.0.

The new algorithm should get the best valid match that is guaranteed to
be of the same compatibility version, but also the most recent version
that is the same or older than the one being installed.

This PR doesn't yet solve the code duplication issue between installAll
and options.