OpenGL Renderer : Add "gl:renderToCurrentContext" command
The original implementation of `render()` didn't match the intended semantics defined by the base class : rendering to the outputs defined by calls to `output()`. It just happened to be implemented in such a way as to be useful to the SceneGadget, rendering into the current framebuffer. We now provide an explicit command for doing this, and `render()` remains unimplemented for now.

Arguably, `render()` should be implemented to render to the outputs defined by `output()` calls, but we don't have a use for that currently. And we _do_ want to ignore outputs destined for other renderers when using an OpenGL render alongside another in a CompoundRenderer. Longer term we might need a mechanism for determining which outputs belong to which renderer, but that can wait until we have a use case.