RenderController : Fix crash changing ArnoldMeshLight filtering
ArnoldMeshLight has a separate bug whereby it inadvertently edits autobump visibility attributes. That will be dealt with in another commit. Here we deal with a bug that those attribute edits exposed in the RenderController.

Autobump attributes cause edit failures in `ArnoldObjectInterface::attributes()` because they need to be applied to the `polymesh` rather than the `ginstance`. This causes the RenderController to drop the current ObjectInterface and create a new one via `RenderController::updateObject()`. But in this case, the object has also switched type from `SceneGraph::LightType` to `SceneGraph::ObjectType`, so `updateObject()` nullifies `m_objectInterface`. This then triggered crashes when we tried to update the transform on the null interface.

The solution is simple : repeat the `if( m_objectInterface )` check after calling `updateObject()` again.