Spreadsheet : Rename `activeRowNames` plug to `enabledRowNames`
We want to preserve the meaning of "active" to be "contributing to the output", in keeping with `Spreadsheet::activeInPlug()` and `Switch::activeInPlug()`. We  are already using "enabled" to mean "considered for matching" in `Row::enabledPlug()`, so it is natural to adopt it for `enabledRowNames`.

The `startup/Gaffer/spreadsheetCompatibility.py` shim provides backwards compatibility for existing scripts.