RenderController : Omit default camera for CompoundRenderer
This stops unwanted cameras appearing at the origin when using alternative viewport renderers. The approach isn't particularly principled, because it relies on the RenderController knowing the details of its usage in SceneGadget. But it works, and it doesn't seem much worse than the assumptions we were already making about pure OpenGL renders.