ParameterHandler : Fix potential array overrun
As discussed in https://github.com/GafferHQ/gaffer/pull/4649, GCC 11 highlighted a potential array overrun in `setupColorPlug()`. We were assuming that RGBA parameters would always be used to create a Color4fPlug, but that is not the case if the plug type has been overridden to Color3fPlug with `gaffer.plugType` metadata. In that case, we could be writing past the end of a Color3f default value, using the alpha from the RGBA parameter default.

I've also refactored the code for getting overrides for the default from the metadata. Although there was no overrun in this part of the code, it was assuming that the metadata size should match the plug size, which is not necessarily the case if one user provides a metadata default and another overrides the plug type. Refactoring allows us to deal with all potential size mismatches in a single well documented call to `copy_n()`.