TweakPlug : Hide `modeToString()`
This was originally implemented in the `.cpp` file, but got moved into a header as part of the functor-based `applyTweaks()` improvement. We still need access to it from the header, so it is now a private static method.

NumericTweak wouldn't have had access to such a private method, but now we require C++17 we can replace it with an inline lambda that does have access. This also removes some boilerplate and makes `applyNumericTweak()` a little easier to follow.