CollectScenes : Don't create empty context variables
Typically we would expect people to use the (non-empty) default for `rootNameVariable`, or to provide their own (non-empty) name. But it is legitimate to use CollectScenes just to collect identical copies of something, or in the absence of direct support in Group, to use it purely to nest something multiple levels deep in the hierarchy. In this case, it natural to want to empty the `rootNameVariable` to remove any additional overhead from the additional context. This commit supports that, by making sure we don't make a bogus variable named `""` in this case.