ImageTestCase : Test pixels outside displayWindow in assertImagesEqual
Previously we incorrectly tested pixels in the display window, instead of the
data window.  Pixels in the display window but outside the data window don't matter,
the Stats node would ignore them anyway.  But pixels in the data window outside
the display window do matter - previously, we failed to check any pixels in the
overscan to see if images are equal.