StringPlug/StringVectorDataPlug : Support interconnection
When driving a StringPlug with a StringVectorDataPlug, we join the array using spaces to form the string value. When driving a StringVectorDataPlug with a StringPlug, we split the string on spaces to form the array value. This is handy in a number of scenarios where we'd previously have resorted to a Python expression to perform the conversion, making the graph harder to understand and slower to execute. My favourite example is using `Spreadsheet.enabledRowNames` to drive `Set.name`, to allow arbitrary numbers of sets to be defined with a single spreadsheet.

We've chosen space as the separator because it fits well with existing usage in Gaffer, where StringPlugs often accept a space-separated list. While other separators - like `,` - are a perfectly reasonable choice, there is no precedent for their use in Gaffer, and we're comfortable with hardcoding a choice which favours existing practice. Other conventions can be supported using conversion expressions as before.

I've made no attempt to quote strings containing spaces when converting from StringVectorDataPlug values, because that quoting would not be supported by any of the nodes consuming the StringPlug value. Typically those nodes either split on spaces or they use `StringAlgo::matchMultiple()`, neither which support quoting.