ImageTransformTest : Raise tolerance on testConcatenation
I played around with this a bit to assess why the threshold is so high ... the worst case match happens to pair a large scale up with a large down, and introduces quite a bit of blur.  This test case looks like it happened to end up completely outside the display window, so we weren't actually testing it at all before.  I guess my conclusion is that it's completely reasonable to just raise the threshold