ie/options : Avoid `moc` wrapper
By making sure that we run the `moc` executable directly, we indicate to
SCons that the generated `moc` file depends on that specific Qt version,
via the resolved filesystem path to the executable itself. If a
different `moc` executable would be called during build, SCons knows
that it needs to regenerate the target.

Previously, we were picking the `moc` wrapper, which had a fixed
filesystem location, regardless of the Qt version being used, which was
causing the moc file to not be regenerated, and the gaffer build to
fail.