TaskNodes : Make virtual overrides protected
Just as the implementation of all ComputeNode overrides is protected, so should be the implementation of all TaskNode overrides. The general design principle is that Nodes provide internal implementation, and Plugs provide public interfaces.

These methods are protected on the TaskNode base class already, and remain exposed to Python as before, so in practice I don't think this change will affect anyone. A bigger change that we should probably consider at some point is to change the Python bindings to prefix the protected methods with `_`. That would make it much clearer that the TaskPlug API is the one that people should be interacting with.