ImageAlgo::parallelProcess/GatherTiles : Change interpretation of "window"
Previously, any empty data window meant "evaluate the data window from the image", but now, only a default data window causes the images data window to be evaluated - a non-default empty data window forces an empty data window to be used.  This new behaviour seems to match the original description "// Uses dataWindow if not specified" - it didn't say "Uses dataWindow if empty".

This is now important because it is legimate to use these functions without a view name scoped, as long as the functor takes responsibility for scoping a view name - but this is not true if you need to evaluate the data window.