GafferCyclesTest : Add InteractiveCyclesRenderTest
The base class implements 29 tests. We need to skip 2 due to as-yet unresolved issues with the output driver. We skip a further 5 due to missing features in Cycles itself. Everything else is a pass.

The relaxed comparison in `testAddLight()` is necessary because there's slight colour noise in the Cycles render. I suspect this is because Cycles chooses a different light to sample per-camera-ray, whereas Arnold etc are sampling all lights for each ray.

It was also necessary to adjust the shader handling logic in CyclesAttributes for `testAddLight()` to pass. The issue was that the surface shader was being assigned to all locations, even the light. The CyclesAttributes code was preferring this over the light shader itself, preventing the light from emitting. We now only consider the `ccl:light` attribute when looking for light shaders.