CI : Remove Mac testing
This does not mean me are permanently giving up on the Mac platform. In fact, we are doing more development on Mac than we have for a long time. But our development machines are all Apple Silicon now, and we have no time or motivation to maintain Intel builds. GitHub Actions doesn't currently provide Apple Silicon workers, so we are removing the CI until that changes, or until we can set up a self-hosted worker.

The latest dependencies release (5.0.0a5) provides packages for Apple Silicon but not for Intel. Of course, if someone volunteers to maintain Intel Mac packages, we will be happy to continue to host and test against them. But to date we have received no support on the Mac front, so it remains what it has always been - our best efforts based on doing some of our personal development on Mac.