RenderController : Fix duplicate callbacks from `updateInBackground()`
When a set of priority paths were passed, we were calling `callback( Completed )` twice. This could cause crashes when the SceneGadget was used with a raytraced renderer. On the first call, we were calling `renderer->render()` to start the render, which meant that when the non-priority paths were edited, the renderer was running (edits are only allowed while paused).