SceneGadget : Fix selection when switching to Arnold renderer
At the point we were calling `OutputBuffer::setSelection()`, the new RenderController had not exported the scene yet, and therefore had not allocated IDs for any objects. We now pass `createIfNecessary = true` to `RenderController::idsForPaths()` so that IDs are pregenerated and the OutputBuffer can draw the selection properly once the render becomes available.