Changes.md : Add missing entry for Viewer crash fix
I was in two minds about including this, because we won't want to include it in the final release notes for `0.62.0.0`. But we do want an accurate record of changes if we put out an `a4` release, so I'm adding it with a footnote to remind me to remove it when consolidating all the changes into one for the final release.