ValuePlug : Fix subtle cache key problem
This fixes a crash introduced by the StringVectorDataPlug->StringPlug connections added in 2e22984398c643ceeead771790a8ad09650f9670. The crashes were the result of infinite recursion between `StringPlug::hash()` and `StringPlug::getObjectValue()`. This is exposed initially by the addition of a StringVectorDataPlug->StringPlug connection in `SetTest.testSetVariable()`, which is where I encountered the problem in the real world.

The underlying issue is that we have been using `ValuePlug::hash()` for two subtley different purposes :

- In ComputeProcessKey, as a key into the cache where we store `IECore::Object` values for plugs.
- In the public API as a hash representing the result of `ValuePlug::getValue()` calls.

For most plugs, these two things are one and the same, because `getValue()` returns an object from the cache directly and the default `ValuePlug::hash()` is sufficient for both. But for StringPlugs, they are different, because `StringPlug::getValue()` performs additional context-based substitutions on top of the cached value. Therefore `StringPlug::hash()` overrides `ValuePlug::hash()` to provide a hash that exactly corresponds to the subsituted value.

Where the StringVectorDataPlug->StringPlug connections come in is that `ValuePlug::hash()` has dedicated logic to account for the conversion :

```
// Get the hash from
// the input and add a little extra something to represent the
// conversion that m_resultPlug->setFrom( input ) will perform,
// to break apart the cache entries.
IECore::MurmurHash h = input->hash();
h.append( input->typeId() );
h.append( plug->typeId() );
return h;
```

This is the cache key we want to store the result of the conversion in the cache. But the `StringPlug::hash()` override skips all of that logic, and instead calls `getObjectValue()`. In turn, `getObjectValue()` tried to retrieve the value from the cache, which meant computing the cache key, which meant calling `StringPlug::hash()`. Which caused infinite recursion and a crash.

This is fixed by recognising that `StringPlug::hash()` is _not_ the appropriate cache key and using the base `ValuePlug::hash()` instead. In future we should perhaps consider changes to the API so that the two concepts are explicitly separated, but it's not immediately clear how, and that would be a breaking change anyway.

I've also deprecated the `precomputedHash` argument to `StringPlug::getValue()`, because it's far too much to expect client code to understand this subtley and to pass the right flavour of hash. In practice, this isn't used anyway so will easily removed in future.

The final detail is why `StringPlugTest.testStringVectorDataInput()` didn't catch this problem in the first place. The reason is that the cache policy for the conversion is chosen by calling `ComputeNode::computeCachePolicy()`, and because there was no ComputeNode in the test, we fell back to using an uncached conversion, where the hash is never needed. Perhaps there is some debate to be had about whether we ever want to cache conversions, but I want to get the existing behaviour _correct_ before considering changes that need measuring for performance.