BleedFill : Temporary fix for multi-view.
The desirable result for bleed-filling an image with multiple views would be to fill each view independently.

The issue with the previous setup however is that it uses the size to drive a number of loop iterations for a merge.  Evaluating the view names of the output requires evaluating the loop, which means evaluating the number of iterations, which means evaluating the format - but you can't do that from a viewNames context which doesn't have the view set.

The quickest solution was to add a FormatQuery fixed to the default view to grab the format.  A better solution that properly supports multiview would be to use CollectViews to make it clear that the loop iterations are being set independently for each view ... or maybe you could hack it just by using a merge to overwrite the view names so the don't get evaluated through the loop?