SceneGadget : Fix error handling bug
The `m_updateErrored` flag tells us if the previous render controller update ended in error. It is needed so we can avoid displaying partial results from the OutputBuffer and the GL Renderer's scene graph. It is _not_ a flag that tells us we shouldn't try to update again - whether or not we need to update is determined entirely by `RenderController::updateRequired()`. In 3feea6a1f7cbb91a85f71ffb18f8efb71e819064 I made the mistake of thinking it was the latter, putting the check _before_ `updateRenderer()` instead of after it. This meant that the SceneGadget would never update again after an initial error.