ImageAlgo : Replace `sortChannelNames()` with `sortedChannelNames()`
This matches the semantics we had implemented for the Python bindings. I don't see any harm in the slight additional overhead for the C++ API, given that the places it is used are not performance-critical. I find it hard to imagine a performance-critical place (a compute?) where we wouldn't be starting from a `const` container anyway (as retrieved from `ImagePlug::channelsNamesPlug()->getValue()`).