ShaderTweaksUI : Allow creation and removal of parameters
ShaderTweaks was originally quite restrictive so that folks could only modify the parameters that were made by the upstream Shader nodes. But for debugging "data-smithing" of incoming USD data, I'm finding Create and Remove to be too useful to omit. If there's any objection to this I could move it to a personal config, but I think it's likely to be useful more widely than that.