CatalogueUI : Catch exceptions during __pathListingDrop
Somehow, if the image write fails ( for example, dropping an image with no views ), and this exception is allowed to escape, this results in something incorrectly holding a reference to _ImageListing.  If you then close and reopen the UI, the old Qt objects have been deleted, but the Python versions have been kept alive.  The next drag and drop of a valid image will cause a segfault when __pathChanged on _ImageListing.__pathListing calls __update, which uses _qtAddress of the Qt object which has been freed.