UVView : Pass context to ImageGadgets used for textures
Without this fix, texture display was broken in Gaffer-1.0 because the default imageView wasn't available.

This would have been much easier to spot if ImageGadget::renderLayer didn't silently catch and discard exceptions ... the problem is that currently we set up ImageGadgets for UDIMs that don't have textures, and we don't want to get a pile of warnings for those, so I'm ignoring that issue for now.