CyclesLight : Fix metadata queries during copy/paste
The code in CyclesShaderUI uses the Shader's `name` plug to look up metadata that determines whether or not nodules are visible for the light's parameters. So we need to set `name` _before_ adding the plugs for the parameters, so that metadata is available when the plugs are created.

A more principled fix might be to signal that the metadata has changed whenever the `name` plug is set. But this would cause a bunch of unnecessary churn during the construction of the NodeGadget, as nodules are added and then removed unnecessarily. We don't expect the `name` to change at any other time than in `loadShader()`, so the simple fix here is the most pragmatic choice.