CyclesOptions : Remove `featureSet` plug
The `experimental` flag appears to do nothing at all in Cycles at present, so there is no point exposing it. We're also tending to think that we don't want to expose it directly anyway, as we don't want to be on the hook for supporting unsupported features. If in future someone needs to turn this on, they can use CustomOptions.