GafferCyclesUITest : Add CyclesSceneGadgetTest
This includes an additional test case for changing the size of the viewport a bunch of times, and also fixes a camera handling bug in SceneGadget. When the SceneGadget wasn't in a ViewportGadget yet, as in `SceneGadgetTest.testBound()`, we were attempting to start a render using a camera we hadn't created yet. We now skip rendering when we don't have a viewport camera.