Cycles Renderer : Disable resolution divider
This stops Cycles from doing a series of low-res renders before starting the final full-res render in interactive mode. This behaviour might actually be useful, but our display driver doesn't currently know about it, so we end up with miniature renders in the corner of the framebuffer. Disabling for now, but can revisit later.