Cycles Renderer : Reset session _before_ setting output driver
As documented in the comment, this stops the Cycles render threads from trying to send tiles to an output driver that has just been deleted on the main thread. This in turn prevents a crash when resizing the Cycles viewport.

I don't think the current implementation of `pause()` meets the expectations of
the Renderer API : that the renderer should have stopped doing anything before it returns. Currently I am convinced that Cycles continues rendering for a bit after we call `Session::set_pause()` in `CyclesRenderer::pause()` - there is no wait for the rendering threads in `Session::set_pause()`.

If we could refactor `pause()` so that it works as expected, the I don't think this workaround would be necessary. At that point, we should be the only one accessing the session and the scene in between `pause()` and `render()`. I have made attempts in this direction by using `Session::cancel()` in `pause()`, and I'm pretty sure this _does_ wait for the renderer to be done. But that seems to cause crashes of its own in the BVH build code and I have no idea why. So until I understand the bigger picture I am going with this far simpler workaround.