PathListingWidget : Fix selection bug
This was revealed in the previous commit fixing the selection of the bottom row of a list. If the user selected a row that was half visible Qt would scroll the to make the widget visible on button press because it was being selected.

On button release, we were again making  a selection, but now that the list has been scrolled in button press, a new row is under the mouse and is selected.

Removing the first selection prevents the auto-scroll and incorrect selection.