SConstruct : Fix stale file substitutions
SCons didn't wasn't noticing changes to the file substitutions, presumably because of the extra level of indirection through the variable substitutions. This led to stale version numbers being baked into headers and the `About` module, which in turn put the wrong version number in the Gaffer title. Performing the substitution ourselves beforehand fixes this.