PathListingWidgetBinding : Add error indicator
Previously, errors in computation of light editor cell values caused warnings on the console but no indication in the UI. See #4805 for more details. A generic handler in PathModel deals sufficiently with most errors, but we also have a custom handler in LocationNameColumn so that we can keep the location name in addition to the error icon.