Rename : Add `useRegularExpressions` plug
When on, this performs find & replace using `std::regex_replace`. I'm adding this semi-reluctantly, because regular expressions are a good example of the kind of user-unfriendly complexity that I'd like to keep out of Gaffer. But I'm convinced that if we don't add them, someone will use a Python expression to do it manually, and that will be even worse.

An alternative might be to extend our own `StringAlgo::match()` code with a capture syntax. But then I imagine we'd also want a way of anchoring matches at the start or the end, and so on, and then we find ourselves back at regexes but with them now polluting the entire application.

Open questions :

- Should we be using the ECMA grammar that is the default, or the basic POSIX grammar, the extended POSIX grammar, the awk grammar, the grep grammar or the egrep grammar? What's the difference anyway? I'm going off adding this already :)
- Should we be using the ECMA replacement syntax (`$1`) for referencing captures, or the POSIX synax (`\1`), or our own (`<1>` perhaps).
- If using `$1` or `\1` syntax for replacements, should we be doing something helpful so that they don't need additional escaping in Gaffer?