Constraint : Add private `affectsTarget()` method
It's easier to validate affects/hash/compute if utility methods such as `target()` have an associated `affects()` - otherwise dependency changes are hidden too easily inside the utility function and not accounted for elsewhere. I say this having repeatedly got this wrong during refactoring, and am adding this in retrospect (and rebasing it to the front of the history so it looks like I knew what I was doing).