Signals::ScopedConnection : Add move construction and assignment
These allow "ownership" of the connection to be transferred to a new ScopedConnection, without the original scuppering us by subsequently disconnecting anyway. Among other things, this allows us to store ScopedConnection in `std::vector` without resizing breaking our connections.

Note that this also implicitly deletes the old copy construction and assignment, whose semantics were unsuitable for a class which "owns" a resource (in this case the connection). While this is a breaking change, I don't expect existing code to be affected because the original `boost::signals::scoped_connection` had got this right and deleted them in the first place.