CreateViews : Fall back to `default` view when `image:viewName` missing
We had attempted to keep such fallbacks out of the image nodes themselves by putting `image:viewName` in the main ScriptNode context and the ImageTestCase context in the following commits :

 - 6874c8f98d8c51053912afb68a36428c8b60bacd
 - a157975656d01d6ffdad1fbd467ee69d08f17d9e

But as `CreateViewsTest.testInputToExpressionDrivingEnabledPlug()` demonstrates, this was not sufficient to maintain compatibility with existing production networks because `ImagePlug::viewNames()` _removes_ the `image:viewName`! This led to confusion, bug reports, and a worrying cargo-culting of ContextVariables nodes setting `image:viewName` under _every_ Merge node in at least one production template.

We also had earlier bug reports for pipeline scripts which were executing their own node networks without the benefit of the default script context provided by 6874c8f98d8c51053912afb68a36428c8b60bacd. So it's clear that our original approach to backwards compatibility was flawed.

Given the behaviour of `viewNames()`, the _only_ mechanism sufficient for true backwards compatibility is to move the fallback into the nodes themselves, so it exists no matter the context. That is what I've done here.

Now, we did consider doing this in the first place, and rejected it for a couple of reasons :

1. We didn't want there to be two contexts that mapped to the same image, as this would double up entries in the hash cache. This is a legitimate concern, but our key image consumers (ImageWriter and Viewer) do explicitly provide `image:viewName` already. We also have the option of tightening this up over time, for instance by adding a check in `GafferImageTest::ContextSanitiser`.
2. It might be confusing if users used `${image:viewName}` in a string substitution (on a Text node, say), and didn't get the expected `default`. This should not occur for our own code (because we should always provide `image:viewName`). If it happens for other setups, it gives us an opportunity to educate and suggest the explicit provision of `image:viewName`.

The only other alternatives I can see are :

1. Abandon backwards compatibility, and tell folks to change their templates. I don't think this is a great option.
2. Don't remove `image:viewName` from the context in `viewNames()`. This leads to logical impossibilities like the available views being varied according to the value of `image:viewName`.

Given all the above, I think the approach taken in this commit is the best option.