PathColumn : Add support for showing different icons when hovered
The approach taken here avoids any event handling on the part of the PathColumn. Instead, the column simply provides two icons, one for each state, and then the PathListingWidget takes care of showing the right one at the right time. The idea being to avoid any overhead of evaluating `PathColumn::cellData()` again based on enter/leave events, so we're as fast as we can be.

I'm not 100% sold on the "CompoundData of icons" approach. Other possibilities are :

- Passing a `state` argument to `PathColumn::cellData()`, and storing two versions of the entire cell state. But the idea of a different tooltip for the highlighted state seemed a bit daft. And it means another call to `PathColumn::cellData()`, and the idea of this approach was to reduce overhead.
- Adding a `highlightedIcon` member to the CellData struct. This is a little more explicit, but given that we've already accepted a slightly weird API with DataPtr for each field, it seemed easier to accept one more data format instead. It also means there are no ABI breaks here at all.

This does have limitations though. The main use of the highlighted state is to indicate that something will happen if you click or whatever. But we can't have different states based on mouse position or keyboard state, which could be useful. For instance, I would like a shift+click in the Catalogue to assign the next output number rather than 1. So maybe in future we'll also need event-handling-based updates to hover state. But even then, I think this cheap-and-cheerful mechanism might still have its place for the simple cases.